Automatically Merged using SST Pull Request AutoTester
PR Title: b'Naming and Warning Cleaning'
PR Author: hughes-c

+++ src/sst/elements/messier/nvm_params.h
@@ -24,215 +24,215 @@
 #include <sst/core/component.h>
 #include <sst/core/timeConverter.h>
 #include <sst/elements/memHierarchy/memEvent.h>
-#include<map>
-#include<list>
-//#include "Messier.h"
+
+#include <map>
+#include <list>
 
 // This class structure represents the NVM parameters
-
 using namespace SST; //::MessierComponent;
 
-namespace SST{ namespace MessierComponent{
-
-//using namespace SST::MessierComponent;
+namespace SST{
+namespace MessierComponent{
 
 class NVM_PARAMS
 {
-	public:
+    public:
 
-		// The size of the whole rank in KB
-		long long int size;
+        // The size of the whole rank in KB
+        uint64_t size;
 
-		// The size of the write buffer in number of entries
-		int write_buffer_size;
+        // The size of the write buffer in number of entries
+        uint32_t write_buffer_size;
 
-		// This determines the number of maximum buffer requests inside the controller
-		int max_requests;
+        // This determines the number of maximum buffer requests inside the controller
+        uint32_t max_requests;
 
-		// The number of possible outstanding requests inside the internal controller (not that once writes are written to the write buffer, they are deleted)
-		int max_outstanding;
+        // The number of possible outstanding requests inside the internal controller (not that once writes are written to the write buffer, they are deleted)
+        uint32_t max_outstanding;
 
-		// For power budget (at any time, the controller enforces write_weight*num_current_writes + read_weight*num_current_reads less than or equal max_current_weight
-		float max_current_weight;
+        // For power budget (at any time, the controller enforces write_weight*num_current_writes + read_weight*num_current_reads less than or equal max_current_weight
+        float max_current_weight;
 
-		// This determines the relative power consumption of a PCM write operation
-		float write_weight;
+        // This determines the relative power consumption of a PCM write operation
+        float write_weight;
 
-		// This determines the relative power consumption of a PCM read operation
-		float read_weight;
+        // This determines the relative power consumption of a PCM read operation
+        float read_weight;
 
-		// The clock of the internal controller of the DIMM in GHz
-		float clock;
+        // The clock of the internal controller of the DIMM in GHz
+        float clock;
 
-		// The clock of the internal PCM memory
-		float memory_clock;
+        // The clock of the internal PCM memory
+        float memory_clock;
 
-		// The clock of the IO bus, this typically should be similar to the internal controller clock
-		float io_clock;
+        // The clock of the IO bus, this typically should be similar to the internal controller clock
+        float io_clock;
 
-		// The time to send a command from the internal controller till buffered inside the NVM chips
-		int tCMD;
+        // The time to send a command from the internal controller till buffered inside the NVM chips
+        uint32_t tCMD;
 
-		// This is the access latency of a column for an already activated row buffer
-		int tCL;
+        // This is the access latency of a column for an already activated row buffer
+        uint32_t tCL;
 
-		// This is the actual read latency of a row and installing it on the row buffer
-		int tRCD;
+        // This is the actual read latency of a row and installing it on the row buffer
+        uint32_t tRCD;
 
-		// The latency of writing a column. Note that we assume internal PCMs write it directly to the PCM cells and update the row buffer (in case it was row-buffer hit)
-		int tCL_W;
+        // The latency of writing a column. Note that we assume internal PCMs write it directly to the PCM cells and update the row buffer (in case it was row-buffer hit)
+        uint32_t tCL_W;
 
-		// The time needed for submitting a data burst on the bus
-		int tBURST;
+        // The time needed for submitting a data burst on the bus
+        uint32_t tBURST;
 
-		// This determines the device width in bits
-		int device_width;
+        // This determines the device width in bits
+        uint32_t device_width;
 
-		// num ranks
-		int num_ranks;
+        // num ranks
+        uint32_t num_ranks;
 
-		// This determines the number of devices on the DIMM Ranks (the DIMM bus width is num_devices X  device_width)
-		int num_devices;
+        // This determines the number of devices on the DIMM Ranks (the DIMM bus width is num_devices X  device_width)
+        uint32_t num_devices;
 
-		// This determines the number of banks on each Rank
-		int num_banks;
+        // This determines the number of banks on each Rank
+        uint32_t num_banks;
 
-		// This determines the row buffer size
-		int row_buffer_size;
+        // This determines the row buffer size
+        uint32_t row_buffer_size;
 
-		// This indicates the threshold for starting the write buffer flushing
-		int flush_th;
+        // This indicates the threshold for starting the write buffer flushing
+        uint32_t flush_th;
 
-		// This indicates the low flush threshold
-		int flush_th_low;
+        // This indicates the low flush threshold
+        uint32_t flush_th_low;
 
-		// This indicates the maximum number of conccurent writes to NVM chips (not including those on the write buffer)
-		int max_writes;
+        // This indicates the maximum number of conccurent writes to NVM chips (not including those on the write buffer)
+        uint32_t max_writes;
 
-		// Determines if cacheline interleaving or bank interleaving
-		bool cacheline_interleaving;
+        // Determines if cacheline interleaving or bank interleaving
+        bool cacheline_interleaving;
 
-		// This implements the adaptive writes policy for NVM devices
-		bool adaptive_writes;
+        // This implements the adaptive writes policy for NVM devices
+        bool adaptive_writes;
 
-		// This indicates if there is enough power to back-up the internal cache
-		bool cache_persistent;
+        // This indicates if there is enough power to back-up the internal cache
+        bool cache_persistent;
 
-		// This indicates if the cache is enabled
-		bool cache_enabled;
+        // This indicates if the cache is enabled
+        bool cache_enabled;
 
-		// This indidicates the cache size in KB
-		long long int cache_size;
+        // This indidicates the cache size in KB
+        uint64_t cache_size;
 
-		// This is the associativity of the internal cache
-		long long int cache_assoc;
+        // This is the associativity of the internal cache
+        uint64_t cache_assoc;
 
-		// This indicates the latency of the cache
-		int cache_latency;
+        // This indicates the latency of the cache
+        uint32_t cache_latency;
 
-		// This indicates the cache block size
-		int cache_bs;
+        // This indicates the cache block size
+        uint32_t cache_bs;
 
-		// This inidicates the size of the group of banks to be locked when writing in case of adaptive writes
-		int group_size;
+        // This inidicates the size of the group of banks to be locked when writing in case of adaptive writes
+        uint32_t group_size;
 
-		// This inidicates the locking period in case of adaptive writes
-		int lock_period;
+        // This inidicates the locking period in case of adaptive writes
+        uint32_t lock_period;
 
-		// This indicates if the module scheduling of reads/writes is used
-		bool modulo;
+        // This indicates if the module scheduling of reads/writes is used
+        bool modulo;
 
-		// This indicates the unit of module, if N, this means for each N reads, we service one write
-		int modulo_unit;
+        // This indicates the unit of module, if N, this means for each N reads, we service one write
+        uint32_t modulo_unit;
 
-		// This indicates if the write cancellation technique is used
-		bool write_cancel;
+        // This indicates if the write cancellation technique is used
+        bool write_cancel;
 
-		// This indicates the write cancellation threshold
-		int write_cancel_th;
+        // This indicates the write cancellation threshold
+        uint32_t write_cancel_th;
 
 
-	public:
+    public:
 
-		void operator = (const NVM_PARAMS &D ) {
+        void operator = (const NVM_PARAMS &D ) {
 
 
-			cache_enabled = D.cache_enabled;
+            cache_enabled = D.cache_enabled;
 
-			cache_size = D.cache_size;
+            cache_size = D.cache_size;
 
-			cache_assoc = D.cache_assoc;
+            cache_assoc = D.cache_assoc;
 
-			cache_latency = D.cache_latency;
+            cache_latency = D.cache_latency;
 
-			cache_bs = D.cache_bs;
+            cache_bs = D.cache_bs;
 
-			size = D.size;  // in KB, which mean 8GB
+            size = D.size;  // in KB, which mean 8GB
 
-			cacheline_interleaving = D.cacheline_interleaving;
+            cacheline_interleaving = D.cacheline_interleaving;
 
-			adaptive_writes = D.adaptive_writes;
+            adaptive_writes = D.adaptive_writes;
 
-			write_buffer_size = D.write_buffer_size;
+            write_buffer_size = D.write_buffer_size;
 
-			max_outstanding = D.max_outstanding;
+            max_outstanding = D.max_outstanding;
 
-			max_current_weight = D.max_current_weight;;
+            max_current_weight = D.max_current_weight;;
 
-			write_weight = D.write_weight;
+            write_weight = D.write_weight;
 
-			read_weight = D.read_weight;
+            read_weight = D.read_weight;
 
-			clock = D.clock;
+            clock = D.clock;
 
-			memory_clock = D.clock;
+            memory_clock = D.clock;
 
-			io_clock = D.io_clock;;
+            io_clock = D.io_clock;;
 
-			tCMD = D.tCMD;
+            tCMD = D.tCMD;
 
-			tCL = D.tCL;
+            tCL = D.tCL;
 
-			tRCD = D.tRCD;
+            tRCD = D.tRCD;
 
-			tCL_W = D.tCL_W;
+            tCL_W = D.tCL_W;
 
-			tBURST = D.tBURST;
+            tBURST = D.tBURST;
 
-			device_width = D.device_width;
+            device_width = D.device_width;
 
-			num_ranks = D.num_ranks;
+            num_ranks = D.num_ranks;
 
-			num_devices = D.num_devices;
+            num_devices = D.num_devices;
 
-			num_banks = D.num_banks;
+            num_banks = D.num_banks;
 
-			row_buffer_size = D.row_buffer_size;
+            row_buffer_size = D.row_buffer_size;
 
-			flush_th = D.flush_th;
+            flush_th = D.flush_th;
 
-			max_requests = D.max_requests;
+            max_requests = D.max_requests;
 
-			max_writes = D.max_writes;
+            max_writes = D.max_writes;
 
-			flush_th_low = D.flush_th_low;
+            flush_th_low = D.flush_th_low;
 
-			cache_persistent = D.cache_persistent;
+            cache_persistent = D.cache_persistent;
 
-			group_size = D.group_size;
+            group_size = D.group_size;
 
-			lock_period = D.lock_period;
+            lock_period = D.lock_period;
 
-			modulo = D.modulo;
+            modulo = D.modulo;
 
-			modulo_unit = D.modulo_unit;
+            modulo_unit = D.modulo_unit;
 
-			write_cancel = D.write_cancel;
+            write_cancel = D.write_cancel;
 
-			write_cancel_th = D.write_cancel_th;
+            write_cancel_th = D.write_cancel_th;
 
-		}
+        }
 };
-}}
+
+}
+}
 
 #endif
